Moves `testAccPreCheck`
